refactor(crash): extract backend-agnostic crash collection (CORE-554)
Everything gathered when the process is dying -- the OBS configuration
zip, the window screenshot, the walked stack, the async- and API-call
context, the system and performance snapshots -- is identical whichever
reporting backend is compiled in. Only the submission differs.

Move that collection out of the BugSplat handler into a new
StreamElementsCrashContext, which walks the crashing thread's stack,
decides the module-of-interest gate, and returns attributes, notes and
attachments as plain UTF-8 data. The BugSplat handler keeps its exception
filter, message window and hang detection, and becomes a translation
layer: UTF-8 in, wide strings out, into setAttribute / setNotes /
sendAdditionalFile.

Behaviour is unchanged. The eight inline BugSplat SDK calls that used to
sit inside AddObsConfigurationFiles() are now one loop each, submitting
in the same order, and the zip is built exactly as before.

Two pre-existing problems surfaced while moving the code. Both are
preserved rather than fixed, so this stays verifiable by parity:

- The zip's obs-studio/crashes/crash.log entry has always been empty.
  The global it was read from was declared and reserved but never
  written -- the callback that filled it was removed and the read site
  left behind. "crashes/" is blacklisted as well, so the zip carries no
  OBS crash text from either source.
- GenerateTimeDateFilename(), delete_oldest_file() and
  write_file_content() were never called, and obs_fe_user_config()
  existed only to feed them. Around 120 lines, removed.

The file's header comment described steps 3 and 5 of a flow that no
longer exists; rewritten to match what the code now does.

Windows only for now: the macOS integrations gather nothing beyond a
single attribute set at startup, so there is nothing to extract there
yet.

Verified by building all three backend selections (bugsplat, sentry,
none) locally on Windows -- each compiles only its own handler, with no
warnings under /WX. This project has no tests.
